refactor(select): make popover contents into subcomponent
